move main.go and db.go to the correct locations
