+++ main.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/B0go/itask-backend/config"
 	"github.com/B0go/itask-backend/controller"
-	"github.com/B0go/itask-backend/main/db"
+	"github.com/B0go/itask-backend/db"
 	"github.com/gin-gonic/gin"
 	_ "github.com/lib/pq"
 )
